+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanel.java
@@ -176,7 +176,7 @@ public class EntityPanel extends JPanel {
   private EntityPanel previousSiblingPanel;
   private EntityPanel nextSiblingPanel;
   private boolean toolbarControls = TOOLBAR_CONTROLS.get();
-  private String controlsComponentConstraints = TOOLBAR_CONTROLS.get() ?
+  private String controlComponentConstraints = TOOLBAR_CONTROLS.get() ?
           CONTROL_TOOLBAR_CONSTRAINTS.get() : CONTROL_PANEL_CONSTRAINTS.get();
   private boolean includeControls = true;
   private boolean includeToggleEditPanelControl = INCLUDE_TOGGLE_EDIT_PANEL_CONTROL.get();
@@ -349,8 +349,8 @@ public class EntityPanel extends JPanel {
   /**
    * @return the controls component layout constraints (BorderLayout constraints)
    */
-  public final String getControlsComponentConstraints() {
-    return controlsComponentConstraints;
+  public final String getControlComponentConstraints() {
+    return controlComponentConstraints;
   }
 
   /**
@@ -372,22 +372,22 @@ public class EntityPanel extends JPanel {
    *
    * etc.
    * </pre>
-   * @param controlsComponentConstraints the controls component layout constraints (BorderLayout constraints)
+   * @param controlComponentConstraints the controls component layout constraints (BorderLayout constraints)
    * @throws IllegalStateException if the panel has been initialized
    * @throws IllegalArgumentException in case the given constraint is not one of BorderLayout.SOUTH, NORTH, EAST or WEST
    */
-  public final void setControlsComponentConstraints(String controlsComponentConstraints) {
+  public final void setControlComponentConstraints(String controlComponentConstraints) {
     checkIfInitialized();
-    switch (requireNonNull(controlsComponentConstraints)) {
+    switch (requireNonNull(controlComponentConstraints)) {
       case BorderLayout.SOUTH:
       case BorderLayout.NORTH:
       case BorderLayout.EAST:
       case BorderLayout.WEST:
         break;
       default:
-        throw new IllegalArgumentException("Controls component constraint must be one of BorderLayout.SOUTH, NORTH, EAST or WEST");
+        throw new IllegalArgumentException("Control component constraints must be one of BorderLayout.SOUTH, NORTH, EAST or WEST");
     }
-    this.controlsComponentConstraints = controlsComponentConstraints;
+    this.controlComponentConstraints = controlComponentConstraints;
   }
 
   /**
@@ -746,14 +746,14 @@ public class EntityPanel extends JPanel {
 
   /**
    * Creates a base panel containing the given edit panel.
-   * The default layout is a {@link FlowLayout} with the alignment depending on the {@link #getControlsComponentConstraints()}.
+   * The default layout is a {@link FlowLayout} with the alignment depending on the {@link #getControlComponentConstraints()}.
    * The resulting panel is added at {@link BorderLayout#CENTER} on the {@link #editControlPanel()}
    * @param editPanel the initialized edit panel
    * @return a base panel for the edit panel
    */
   protected JPanel createEditBasePanel(EntityEditPanel editPanel) {
-    int alignment = controlsComponentConstraints.equals(BorderLayout.SOUTH) ||
-            controlsComponentConstraints.equals(BorderLayout.NORTH) ? FlowLayout.CENTER : FlowLayout.LEADING;
+    int alignment = controlComponentConstraints.equals(BorderLayout.SOUTH) ||
+            controlComponentConstraints.equals(BorderLayout.NORTH) ? FlowLayout.CENTER : FlowLayout.LEADING;
 
     return panel(new FlowLayout(alignment, 0, 0))
             .add(editPanel)
@@ -765,18 +765,18 @@ public class EntityPanel extends JPanel {
    * such as insert, update, delete, clear and refresh.
    * Only called if {@link #isIncludeControls()} returns true.
    * @param controls the controls to display on the component
-   * @return the component containing the edit and table panel controls
+   * @return the component containing the edit and table panel controls, null if no controls are available
    * @see EntityEditPanel#controls()
    * @see #createControls()
    * @see EntityPanel#TOOLBAR_CONTROLS
    * @see EntityPanel#CONTROL_PANEL_CONSTRAINTS
    * @see EntityPanel#CONTROL_TOOLBAR_CONSTRAINTS
    */
-  protected JComponent createControlsComponent(Controls controls) {
+  protected JComponent createControlComponent(Controls controls) {
     if (requireNonNull(controls).empty()) {
       return null;
     }
-    boolean horizontalLayout = controlsComponentConstraints.equals(BorderLayout.SOUTH) || controlsComponentConstraints.equals(BorderLayout.NORTH);
+    boolean horizontalLayout = controlComponentConstraints.equals(BorderLayout.SOUTH) || controlComponentConstraints.equals(BorderLayout.NORTH);
     if (toolbarControls) {
       return toolBar(controls)
               .orientation(horizontalLayout ? HORIZONTAL : VERTICAL)
@@ -806,7 +806,7 @@ public class EntityPanel extends JPanel {
    * By default all controls from {@link EntityEditPanel#controls} are included and if a
    * table panel is available a table refresh controls is included as well.
    * @return the control component controls, an empty {@link Controls} instance in case of no controls.
-   * @see #createControlsComponent(Controls)
+   * @see #createControlComponent(Controls)
    */
   protected Controls createControls() {
     Controls controls = Controls.controls();
@@ -993,9 +993,9 @@ public class EntityPanel extends JPanel {
     editControlPanel.setBorder(createEmptyBorder(gap, 0, gap, 0));
     editControlPanel.add(createEditBasePanel(editPanel), BorderLayout.CENTER);
     if (includeControls) {
-      JComponent controlsComponent = createControlsComponent(createControls());
-      if (controlsComponent != null) {
-        editControlPanel.add(controlsComponent, controlsComponentConstraints);
+      JComponent controlComponent = createControlComponent(createControls());
+      if (controlComponent != null) {
+        editControlPanel.add(controlComponent, controlComponentConstraints);
       }
     }
   }