+++ changelog.md
@@ -5,6 +5,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityPanel bug fixed, table refresh control was always enabled, instead of only when the panel was active.
 - EntityPanel.createControls() added, related refactoring.
+- EntityPanel.createControlsComponent() renamed createControlComponent().
 
 ## 0.17.20
 ### is.codion.common.model